prism-pattern/examples/gemini-implmentation/requirements.txt
google-cloud-aiplatform>=1.70.0
pdf2image>=1.17.0
Pillow>=10.1.0
python-dotenv>=1.0.0
